Internals: change Draft-on format to handle all of branch/detached/orphan correctly
With help from https://stackoverflow.com/a/69935596/1979340, properly handle all of
the different states that HEAD can be in without resorting to (too many) hacks.

BREAKING CHANGE: the values of the Draft-on trailer change; versions which emit the
old format do not understand the new and vice versa. The new version has an error
message explaining this situation; the fix is to explicitly pass an --onto option
when switching to drafts created using the old-style trailer.